# artifacts/life-os/lib/api-zod/src/generated/types/routine.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LifeOS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoutineFrequency } from "./routineFrequency";

export interface Routine {
  id: number;
  title: string;
  description?: string | null;
  time?: string | null;
  emoji?: string | null;
  streak: number;
  completedToday: boolean;
  frequency: RoutineFrequency;
  createdAt: string;
}