# artifacts/life-os/lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LifeOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  tasksCompleted: number;
  tasksTotal: number;
  goalsInProgress: number;
  routinesStreak: number;
  journalEntries: number;
  routinesCompletedToday: number;
  routinesTotal: number;
}