# artifacts/life-os/lib/api-zod/src/generated/types/updateGoal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LifeOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateGoal {
  title?: string;
  description?: string | null;
  /**
   * @minimum 0
   * @maximum 100
   */
  progress?: number;
  targetDate?: string | null;
  category?: string | null;
  color?: string | null;
  emoji?: string | null;
}