# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LifeOS API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
  - name: tasks
    description: Task management
  - name: goals
    description: Goal tracking
  - name: routines
    description: Daily routine management
  - name: journal
    description: Journal entries
paths:
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect on success or failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: Get all tasks
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTask"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTask"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Task deleted

  /goals:
    get:
      operationId: getGoals
      tags: [goals]
      summary: Get all goals
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoal"
      responses:
        "201":
          description: Goal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"

  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoal"
      responses:
        "200":
          description: Goal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Goal deleted

  /routines:
    get:
      operationId: getRoutines
      tags: [routines]
      summary: Get all routines
      responses:
        "200":
          description: List of routines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Routine"
    post:
      operationId: createRoutine
      tags: [routines]
      summary: Create a routine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoutine"
      responses:
        "201":
          description: Routine created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Routine"

  /routines/{id}:
    patch:
      operationId: updateRoutine
      tags: [routines]
      summary: Update a routine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoutine"
      responses:
        "200":
          description: Routine updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Routine"
    delete:
      operationId: deleteRoutine
      tags: [routines]
      summary: Delete a routine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Routine deleted

  /journal:
    get:
      operationId: getJournalEntries
      tags: [journal]
      summary: Get all journal entries
      responses:
        "200":
          description: List of journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      operationId: createJournalEntry
      tags: [journal]
      summary: Create a journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJournalEntry"
      responses:
        "201":
          description: Journal entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"

  /journal/{id}:
    patch:
      operationId: updateJournalEntry
      tags: [journal]
      summary: Update a journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJournalEntry"
      responses:
        "200":
          description: Journal entry updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
    delete:
      operationId: deleteJournalEntry
      tags: [journal]
      summary: Delete a journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Journal entry deleted

  /stats:
    get:
      operationId: getStats
      tags: [health]
      summary: Get user stats summary
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        completed:
          type: boolean
        priority:
          type: string
          enum: [low, medium, high]
        category:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - title
        - completed
        - priority
        - createdAt

    CreateTask:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        priority:
          type: string
          enum: [low, medium, high]
        category:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
      required:
        - title
        - priority

    UpdateTask:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        completed:
          type: boolean
        priority:
          type: string
          enum: [low, medium, high]
        category:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true

    Goal:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        progress:
          type: integer
          minimum: 0
          maximum: 100
        targetDate:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        emoji:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - title
        - progress
        - createdAt

    CreateGoal:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        targetDate:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        emoji:
          type: string
          nullable: true
      required:
        - title

    UpdateGoal:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        progress:
          type: integer
          minimum: 0
          maximum: 100
        targetDate:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        emoji:
          type: string
          nullable: true

    Routine:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        time:
          type: string
          nullable: true
        emoji:
          type: string
          nullable: true
        streak:
          type: integer
        completedToday:
          type: boolean
        frequency:
          type: string
          enum: [daily, weekdays, weekends, custom]
        createdAt:
          type: string
      required:
        - id
        - title
        - streak
        - completedToday
        - frequency
        - createdAt

    CreateRoutine:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        time:
          type: string
          nullable: true
        emoji:
          type: string
          nullable: true
        frequency:
          type: string
          enum: [daily, weekdays, weekends, custom]
      required:
        - title
        - frequency

    UpdateRoutine:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        time:
          type: string
          nullable: true
        emoji:
          type: string
          nullable: true
        streak:
          type: integer
        completedToday:
          type: boolean
        frequency:
          type: string
          enum: [daily, weekdays, weekends, custom]

    JournalEntry:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
          nullable: true
        content:
          type: string
        mood:
          type: string
          enum: [great, good, okay, bad, awful]
          nullable: true
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - content
        - tags
        - createdAt

    CreateJournalEntry:
      type: object
      properties:
        title:
          type: string
          nullable: true
        content:
          type: string
        mood:
          type: string
          enum: [great, good, okay, bad, awful]
          nullable: true
        tags:
          type: array
          items:
            type: string
      required:
        - content
        - tags

    UpdateJournalEntry:
      type: object
      properties:
        title:
          type: string
          nullable: true
        content:
          type: string
        mood:
          type: string
          enum: [great, good, okay, bad, awful]
          nullable: true
        tags:
          type: array
          items:
            type: string

    Stats:
      type: object
      properties:
        tasksCompleted:
          type: integer
        tasksTotal:
          type: integer
        goalsInProgress:
          type: integer
        routinesStreak:
          type: integer
        journalEntries:
          type: integer
        routinesCompletedToday:
          type: integer
        routinesTotal:
          type: integer
      required:
        - tasksCompleted
        - tasksTotal
        - goalsInProgress
        - routinesStreak
        - journalEntries
        - routinesCompletedToday
        - routinesTotal